refactor: PictureTag 모듈 생성
