+++ src/main/java/com/onebyte/life4cut/picture/service/PictureService.java
@@ -16,7 +16,7 @@ import com.onebyte.life4cut.picture.domain.PictureTag;
 import com.onebyte.life4cut.picture.domain.PictureTagRelation;
 import com.onebyte.life4cut.picture.repository.PictureRepository;
 import com.onebyte.life4cut.picture.repository.PictureTagRelationRepository;
-import com.onebyte.life4cut.picture.repository.PictureTagRepository;
+import com.onebyte.life4cut.pictureTag.repository.PictureTagRepository;
 import com.onebyte.life4cut.support.fileUpload.FileUploadResponse;
 import com.onebyte.life4cut.support.fileUpload.FileUploader;
 import com.onebyte.life4cut.support.fileUpload.MultipartFileUploadRequest;
